backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
django-storages>=1.14
boto3>=1.26
gunicorn>=21.0
python-dotenv>=1.0
psycopg2-binary>=2.9